frontend

+++ db/migrate/20171031183006_create_historical_rates.rb
@@ -5,8 +5,8 @@ class CreateHistoricalRates < ActiveRecord::Migration[5.1]
     create_table :historical_rates do |t|
       t.integer :source_currency
       t.integer :target_currency
-      t.integer :rate
-      t.date    :date
+      t.float :rate
+      t.date :date
 
       t.timestamps
     end
